pyproject.toml
[project]
name = "academic_dreamer"
version = "0.1.0"
description = "Multi-agent system for creating academic illustrations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "AcademicDreamer", email = "academic@example.com" }
]
keywords = ["academic", "illustration", "diagram", "langgraph", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langgraph>=0.0.20",
    "openai>=1.12.0",
    "pillow>=10.2.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
academic-dreamer = "academic_dreamer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["academic_dreamer"]
include = [
    "/academic_dreamer/prompts",
    "/academic_dreamer/config",
]

[tool.hatch.build.targets.sdist]
include = [
    "/academic_dreamer",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
